Define COEP:cors-or-credentialless
Specify the behavior of `Cross-Origin-Embedder-Policy: cors-or-credentialless`,
Originally described in:
https://github.com/mikewest/credentiallessness

For the HTML spec:
- `COEP:cors-or-credentialless`
- `COEP:require-corp`
have equivalent behavior..

Those value are said to be `compatible with crossOriginIsolation`. This patch
rewrite the specification to use this concept.

This patch is accompagned with a fetch specification patch:
XXX

COEP:cors-or-credentialless causes cross-origin no-cors request to omit
credentials. In exchange, Cross-Origin-Resource-Policy in response isn't
required.